moving editorFunc to Editor.jsx
