+++ frontend/src/Editor/index.js
@@ -2,4 +2,3 @@
 export * from './Editor';
 export * from './Viewer';
 export * from './DataSourceManager';
-export * from './EditorFunc';
