+++ frontend/src/_helpers/appUtils.js
@@ -1471,6 +1471,35 @@ export const cloneComponents = (
   });
 };
 
+// const getAllChildComponents = (allComponents, parentId) => {
+//   const childComponents = [];
+
+//   Object.keys(allComponents).forEach((componentId) => {
+//     const componentParentId = allComponents[componentId].component?.parent;
+
+//     const isParentTabORCalendar =
+//       allComponents[parentId]?.component?.component === 'Tabs' ||
+//       allComponents[parentId]?.component?.component === 'Calendar';
+
+//     if (componentParentId && isParentTabORCalendar) {
+//       const childComponent = allComponents[componentId];
+//       const childTabId = componentParentId.split('-').at(-1);
+//       if (componentParentId === `${parentId}-${childTabId}`) {
+//         childComponent.componentId = componentId;
+//         childComponents.push(childComponent);
+//       }
+//     }
+
+//     if (componentParentId === parentId) {
+//       const childComponent = allComponents[componentId];
+//       childComponent.componentId = componentId;
+//       childComponents.push(childComponent);
+//     }
+//   });
+
+//   return childComponents;
+// };
+
 const getAllChildComponents = (allComponents, parentId) => {
   const childComponents = [];
 
@@ -1487,6 +1516,10 @@ const getAllChildComponents = (allComponents, parentId) => {
       if (componentParentId === `${parentId}-${childTabId}`) {
         childComponent.componentId = componentId;
         childComponents.push(childComponent);
+
+        // Recursively find children of the current child component
+        const childrenOfChild = getAllChildComponents(allComponents, componentId);
+        childComponents.push(...childrenOfChild);
       }
     }
 
@@ -1494,6 +1527,10 @@ const getAllChildComponents = (allComponents, parentId) => {
       const childComponent = allComponents[componentId];
       childComponent.componentId = componentId;
       childComponents.push(childComponent);
+
+      // Recursively find children of the current child component
+      const childrenOfChild = getAllChildComponents(allComponents, componentId);
+      childComponents.push(...childrenOfChild);
     }
   });
 
@@ -1598,6 +1635,8 @@ export const addComponents = (
     };
 
     finalComponents[newComponentId] = newComponent;
+
+    // const doesComponentHaveChildren = getAllChildComponents
   });
 
   if (currentPageId === pageId) {
